requirements.txt
selenium==4.14.0
urllib3==1.26.16
trio==0.22.2
trio-websocket==0.11.1
certifi==2023.5.7
attrs==23.1.0
sortedcontainers==2.4.0
idna==3.4
outcome==1.3.0
sniffio==1.3.0
cffi==1.16.0
wsproto==1.2.0
PySocks==1.7.1
pycparser==2.21
h11==0.14.0
pyyaml==6.0.1
mnemonic
pyperclip
